U3d-b: the wire cluster, and the baseline that is a view rather than a world
ADR-022's whole wire half in one fail-closed schema bump, plus
PROTOCOL_VERSION 3 -> 4: `WireType::Snapshot`'s payload stopped being
opaque game bytes, so a build that predates this would read a
`DeltaHeader` as the game's old snapshot header and find a plausible
tick with everything after it shifted.

R19 is closed. The replication cliff is removed rather than mitigated:
a tick fills a byte budget from the game's ranked list, delta-encodes
against the view **as sent** at the client's acked tick, and packs the
result into as many datagrams as it takes. A client with no usable
baseline gets a keyframe on the `Bulk` stream; a grid over budget
produces a partial view with an honest `culledCount`, never silence.
ADR-018 D3's shared-grid gate lifts with it.

The envelope changed owner, and that is the shape of the slice.
GameLogic keeps the entity record's *meaning* and the tick *tail* (the
order records and the session's high-water mark, opaque exactly as a
summary frame is). The delta header, the records, the ring, the parts
and the keyframe are the engine's, because interest and delta are the
session host's job and `EntityRecord` was always NeuronCore's type.
`Simulation::WriteSnapshot` is gone; the seam is `RankRelevance` +
`WriteEntities` + `WriteTickTail`.

Two things the ADR does not name had to be decided. `RankRelevance`
returns how many of the front are **guaranteed** -- the engine cannot
work that out, and §5c is unimplementable without it. And `ViewFocus`
had to be invented: §4's query is a focus, an extent and a selection,
and §5a's guarantee cannot be kept for a selection nobody told the
server about. That amends ADR-016 §7's "the server has no business
holding this"; the simulation still learns none of it.

`ShipId` and `EntityRecord::id` widened together, because Ids.h's reason
for the first being u16 was that it matched the second. The registry's
durable-load duplicate check stopped being a bitset indexed by id (free
at 64k bits, half a gigabyte at u32) and became a sort. `EntityId` was
introduced as a named alias, because the client's seams passed entity
ids and grid ids side by side as bare u16 and the compiler had nothing
to say when one widened.

Three defects fell out, all older than this slice. `OrderSubmitBytes`
still charged two bytes per ship id after the writer had widened to
four. `ClientApp::BeginContextAction` passed INVALID_ENTITY_ID as
`PickPoint`'s *radius in metres* -- a 65 km pick floor, so the comment
promising "the same pick the selection uses" was not true; the widening
turned that into 4.29 million, which is how it was found. And QUIC does
not put a stream on the wire until something is written to it, so the
client's bulk stream -- opened, started, and silent because the server
speaks first there -- was a stream the server had never heard of, and
every keyframe was refused.

Documentation: ADR-003/004/005/013/014/016/017/018/022, the dependency
map, the risk register, the plan of record, both build orders and the
README all updated. Zero broken links; no doc names a symbol this slice
removed without saying it is history.

Verified: Debug and Release x64 build clean and warning-free.
1,054 tests pass on both (NeuronCore 76, GameLogic 401, NeuronClient
531, NeuronServer 46). The headless self test passes end to end and its
counters show the intended steady state: 3 keyframes over 124 ticks,
one per join, deltas after.

+++ GameLogic/OrderMessages.cpp
@@ -52,7 +52,9 @@ bool WriteOrderSubmit(const OrderSubmit& _order, Neuron::ByteWriter& _writer) no
   _writer.WriteUInt16(_order.shipCount);
   for (std::uint16_t index = 0; index < _order.shipCount; ++index)
   {
-    _writer.WriteUInt16(_order.shipIds[index]);
+    // u32 as of U3d-b: these are the ids the client read off a snapshot,
+    // and the record's id widened with ADR-022 §8a.
+    _writer.WriteUInt32(_order.shipIds[index]);
   }
   _writer.WriteInt32(_order.target.xCm);
   _writer.WriteInt32(_order.target.yCm);
@@ -88,7 +90,7 @@ bool ReadOrderSubmit(Neuron::ByteReader& _reader, OrderSubmit& _outOrder) noexce
   _outOrder.shipCount = shipCount;
   for (std::uint16_t index = 0; index < shipCount; ++index)
   {
-    _outOrder.shipIds[index] = _reader.ReadUInt16();
+    _outOrder.shipIds[index] = _reader.ReadUInt32();
   }
 
   _outOrder.target.xCm = _reader.ReadInt32();
